# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DEMS API",
    "description": "Plataforma de evaluación de prototipos IPN-DEMS.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/admin/editions/{id}": {
      "get": {
        "tags": [
          "admin/editions"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditionView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin/editions"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Tiene rúbricas o prototipos"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin/editions"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditionView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/admin/prototipos/{id}": {
      "get": {
        "tags": [
          "admin/prototipos"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototipoView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin/prototipos"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Tiene evaluaciones"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin/prototipos"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchPrototipoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototipoView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/admin/rubric-templates": {
      "get": {
        "tags": [
          "admin/rubrics"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "edition_id",
            "in": "query",
            "description": "Filtrar por edición",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "description": "Filtrar por tipo",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RubricType"
            }
          },
          {
            "name": "activo",
            "in": "query",
            "description": "Filtrar por activo",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RubricTemplateSummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin/rubrics"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRubricRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricTemplateView"
                }
              }
            }
          },
          "422": {
            "description": "Validación falló"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/admin/rubric-templates/{id}/structure": {
      "put": {
        "tags": [
          "admin/rubrics"
        ],
        "operationId": "replace_structure",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceStructureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricTemplateView"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Edición fuera de preparación — estructura congelada"
          },
          "422": {
            "description": "Validación falló"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditionView": {
        "type": "object",
        "required": [
          "id",
          "year",
          "name",
          "active",
          "phase",
          "created_at"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/EditionPhase"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PatchEditionRequest": {
        "type": "object",
        "properties": {
          "active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PatchPrototipoRequest": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": [
              "string",
              "null"
            ]
          },
          "eje_transversal": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "nombre": {
            "type": [
              "string",
              "null"
            ]
          },
          "plantel": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PrototipoView": {
        "type": "object",
        "required": [
          "id",
          "edition_id",
          "folio",
          "nombre",
          "eje_transversal",
          "categorias",
          "integrantes",
          "created_at"
        ],
        "properties": {
          "categorias": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "descripcion": {
            "type": [
              "string",
              "null"
            ]
          },
          "edition_id": {
            "type": "string",
            "format": "uuid"
          },
          "eje_transversal": {
            "type": "boolean"
          },
          "folio": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "integrantes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegranteView"
            }
          },
          "nombre": {
            "type": "string"
          },
          "plantel": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RubricType": {
        "type": "string",
        "enum": [
          "exhibicion",
          "memoria"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}